--- com.prgrms/src/customer_blacklist.csv
@@ -1,40 +1,0 @@
------------------------------------
-customer_Id: f065331b-b419-47b1-b1bb-92c3b53ab1a2
-customer_name: Ms. Lauren Ondricka
-registered_date: 2022-11-11T20:21:14.167475
------------------------------------
-customer_Id: 7bffa4f4-b1b8-49dd-a205-ec2374900e7c
-customer_name: Ophelia Koelpin
-registered_date: 2022-11-11T20:21:14.172722
------------------------------------
-customer_Id: 97bb93f1-b236-4ba8-8164-ed9ed8e2dd57
-customer_name: Barton Wiegand Jr.
-registered_date: 2022-11-11T20:21:14.173098
------------------------------------
-customer_Id: f35fa905-e4fc-4ed5-be03-5e49821fb6d5
-customer_name: Fabiola Sporer II
-registered_date: 2022-11-11T20:21:14.173382
------------------------------------
-customer_Id: 5f6e5ca3-8be0-4409-898e-55c7548c154b
-customer_name: Magali Halvorson
-registered_date: 2022-11-11T20:21:14.173628
------------------------------------
-customer_Id: 9eca1284-8d78-47b7-be6e-adccbc231de1
-customer_name: Leanne Lynch
-registered_date: 2022-11-11T20:21:14.174468
------------------------------------
-customer_Id: 27b173d2-4ac2-43c9-8e2c-aed7055d2ece
-customer_name: Dr. Melisa Fay
-registered_date: 2022-11-11T20:21:14.174811
------------------------------------
-customer_Id: 07f83723-c5a2-4b45-8b0c-d2584545ec2b
-customer_name: Alta Hane
-registered_date: 2022-11-11T20:21:14.175030
------------------------------------
-customer_Id: c876cf08-f5d3-4894-af66-a787bd7d97aa
-customer_name: Candido Ledner
-registered_date: 2022-11-11T20:21:14.175292
------------------------------------
-customer_Id: 2b5081e3-917d-4196-bb2b-19a660ec6210
-customer_name: Irving Hessel DVM
-registered_date: 2022-11-11T20:21:14.175711
