--- com.prgrms/src/main/resources/application.yml
@@ -1,24 +1,0 @@
-
-file:
-  customer:
-    file-path: src/customer_blacklist.csv
-
----
-
-spring:
-  config:
-    activate:
-      on-profile: dev
-
----
-
-spring:
-  config:
-    activate:
-      on-profile: local
-file:
-  voucher:
-    file-path: src/voucher_memory.csv
-
-
-
