+++ src/main/java/org.prgrms/exception/PaymentCannotBeNegativeException.java
