+++ src/main/java/org.prgrms/voucher/MenuType.java
