+++ src/main/java/org.prgrms/voucher/discountType/Amount.java
