+++ src/main/java/org.prgrms/voucher/discountType/DiscountAmount.java
