+++ src/main/java/org.prgrms/voucher/discountType/DiscountRate.java
